gui: hierarchical dispatch
This mostly backward-compatible change

* removes the global event module dependency from the gui module
* eliminates all non-test-suite abuse of private registry variables
* relys entirely on the event for all registration with factorio.
* separates and modularizes the registries for both event and gui
* fixes all tests to pass
* creates a more flexiable API, mostly while preservering
  compatibility:
  * duplicate listeners are supported in gui now like in event
  * bulk deregistration is supported and documented

Some re-ordering of events will occur.  In particular, if it was
previously possible to set up event registration in such a way that the
event ordering was explicitly interleaved between event-module consumers
and gui-module consumers, this is no longer so.  For a given gui
event_id, all of the gui-module consumers will recieve their events at
the same time.  To such granular notification orderings, it
would now be necessary to revert to using the event module.

Probably the impact of this would be quite minor for most mods but
to be sure, it could concievably be catastrophic for the rare highly
event-order-sensitive mod.