event_spec: flesh out faketorio simulator
We already have a kind of faketorio but it's full of dummies
and now we want it to actually generate some events.  This
makes the test harness more confusing but allows us to test
for desired results rather than peeking into the
incidental data-structures the event module used.

This does not fix any actual tests but simply creates a more
robust harness which should allow the event module (or any test function)
to register for faketorio events, the harness to fire them, and the
expected simulated callbacks to actually occur.

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>