spec: Use dots when requiring 3rd party libraries
Although stdlib has moved from slashes to dots, the rest of the world
has not, which means that the objects we are using in the test suite are
not the same objects the test suite is using internally (which is full
of require('foo.bar')).

This fixes that.  It was the only 3rd party library I could find but if
any more are identified, here is the bash script I used to
fix it:

  for lib in luassert; do
    for f in $(find . -name '*.lua' | xargs grep -l "${lib}"\\/); do
      sed -e '/require.*'"${lib}"'\//s|\([[:alnum:]]*\)/|\1.|g' -i "${f}"
    done
  done

The script is heuristic and quite capable of false positives so check
results first!

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>