event_spec: insulate globals unconditionally
It seems a bit wonky to rely on the default, but override-able describe()
behavior to insulate globals between modules.  So until we decide there is
some utilitarian reason to make isolation optional, I think we should
control it explicitly (because we already know from experience it's
to break and generally sow confusion).

So move the global requires into the outermost describe block and upgrade
that block to an explicit insulate block which will not be overridden
by busted command-line arguments.