Is.Assert: lambda message support
It sometimes arises that we want to do something like:

	Is.Assert(foo, 'bar' .. something_expensive() .. 'baz')

But this code kind-of sucks because we are calculating the value of

	'bar' .. something_expensive() .. 'baz'

unconditionally, whenever this code is run.  This leaves the user
with an awkward choice between not using Is.Assert, various inelegant
workarounds, and simply tolerating the performance impact of such an
invocation.

This enhancement to Is.Assert enables the caller to instead write:

	Is.Assert(foo, function()
		return 'bar' .. something_expensive() .. 'baz'
	end)

In which case, the

	'bar' .. something_expensive() .. 'baz'

lambda code will only run if (not foo) -- that is, when the Assert
fails.

The provided lambda is called without argument and run using xpcall to
ensure that any wierd double-fault situation is appropriately reported.

Also adds corresponding test of this feature, even though, in my tree,
there is no test of the normal Is.Assert(foo, 'bar') case (hopefully,
this will be corrected in some other commit).

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>