LinkedList: drop meta-pairs entirely for now
I suppose this doesn't serve much purpose, except to prevent access to the
instance table, which is not really intended.  Unless we want to make it
work like pairs(), returning the ipairs() items /and/ the raw items (and
perhaps we should...), leave it alone, so that it will provide the raw
items.

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>