Change Event.register to not accept nil handlers, add Event.remove_all.
This change makes it so that Event.register will no longer accept nil handlers, preventing a typo from having the effects of 1) failing to register a handler, and 2) deregistering all associated handlers.  I'm betting that most of the use of Event.register(...) is to actually register a handler, and not unregister all of them.

In order to accomodate unregistering all handlers, Event.remove_all(...) has been added.